# Skip feeding crates into dataflow if dep graph is disjoint from the ones queried