Not as much of a reduction in peak memory usage as I hoped. Good performance improvement though.

**Before:**

```console
load time: 17.91s
dataflow time: 117.53s

70357390 allocations, total 214.5 GB, peak 39.7 GB
```

**After:**

```console
load time: 17.98s
simplification time: 2.58s
dataflow time: 99.19s

65690119 allocations, total 202.2 GB, peak 37.5 GB
```